refactor(tests): move system scaffolding to system/support/ (#26)
