+++ tests/system/support/include/asset_factory.hpp
