+++ tests/system/support/include/binary_asset_factory.hpp
