+++ tests/system/support/include/bitmap_asset_factory.hpp
