+++ tests/system/support/include/jpeg_asset_factory.hpp
