+++ tests/system/support/include/png_asset_factory.hpp
