+++ tests/system/support/include/text_asset_factory.hpp
