+++ tests/system/support/src/binary_asset_factory.cpp
@@ -1,5 +1,5 @@
 #include "../include/binary_asset_factory.hpp"
-#include "../include/file_write_utils.hpp"
+#include "file_write_utils.hpp"
 
 fs::path BinaryAssetFactory::make_valid(const fs::path& dir) const {
     fs::path p = dir / "valid.bin";
